+++ src/services/no-op.ts
@@ -3,7 +3,7 @@ import { Composable, preload } from '../models/composable.js'
 
 export class NoOp implements Composable {
 
-    @preload('no.op', 10)
+    @preload('no.op', 50)
     initialize(): Composable { 
         return this;
     }